Move `Route` to its own file
